build: push docs and skip build

+++ pipeline/cleanup.sh
@@ -1,3 +1,5 @@
 #!/bin/bash -e
-rm ./.npmrc
+if [ -f ./.npmrc ]; then
+    rm ./.npmrc
+fi
 unset NPM_CONFIG_USERCONFIG
